Remove support for GitHub API token authentication
It's dangerous to expose even part of the GitHub API with a centralized token;
people would be able to abuse the mirror and perform requests on behalf of the
account behind the token.

Since composer-velocita now automatically rewrites api.github.com requests to
codeload.github.com, we no longer need this.