Proxy `packages.json` with a single slash to upstream
We matched on `/packages.json` in the location regex and passed that
match as part of the upstream URL, causing the forward slash to be
duplicated. Packagist did not mind this, but the Drupal 8 repository
definitely took offense.

Simplify this by stop using the regex, and explicitly providing the path
to `proxy_pass`. This also means that we no longer support
`packages.json` files in other locations than the root, but as far as I
know, the only file that matters lives in the root.

Fixes #7.