Add `Accept-Encoding` header to cache key
Most webservers respond with `Vary: Accept-Encoding` when presented with
the  `Accept-Encoding` header, which causes Nginx to automatically
include the request header as part of the cache key. Unfortunately,
issue #8 shows that at least GitLab does not respond with the expected
`Vary` header causing the cache key to be too broad.

For example, if you sent a specific request to Velocita with
`Accept-Encoding` set to `br`, the Brotli-encoded response would be
cached. If you would then perform the same request _without_ the
`Accept-Encoding` header, you would erroneously receive that same
Brotli-encoded response.

Fix this by always including the value of the `Accept-Encoding` request
header in the cache key.